Step R69: OwnedQuery moved to Tier B, and the six row-limiting reds needed no new knob (#56)
212 tests, 210 green. `failed` 102 -> 104 and `total` 28964 -> 28982, which is
212 on Tier B less the 194 Tier A tests this move deletes.

Both reds are one #60 theory, dying on the fixture's `RelationalTestStore`
cast; that cast is on no other route, so ADR-013's amendment adopts the base.

The handoff asked whether the server should inherit the fixture's warning
configuration. It should not, and C55 had already measured the global form at
8 fixed / 626 broken. What closes the six row-limiting reds is C69's
per-fixture knob, `AssociationsWarnings.ThrowOnUnorderedRowLimiting`, which
already forwards exactly this event for four Associations fixtures; this is
the fifth caller. No "no throw" override was written -- that would have hidden
a real difference between this provider and every relational one.

The re-parent deletes five InMemory-limitation overrides and a duplicated
TestSqlLoggerFactory. D6 checked and does not apply: no transaction strategy.

Measured r69-ownedquery against r68-noclienteval: FIXED none, BROKEN exactly
the two, one reason moved.
