+++ test/InfoCarrier.Core.FunctionalTests/InMemory/Query/SharedTypeQueryInfoCarrierTest.cs
@@ -1,0 +1,46 @@
+﻿// Licensed under the MIT license. See license.txt file in the project root for license information.
+
+using InfoCarrier.Core.FunctionalTests.TestUtilities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.TestUtilities;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace InfoCarrier.Core.FunctionalTests.InMemory.Query;
+
+/// <summary>
+///     <c>SharedTypeQueryTestBase</c> on Tier A — a model whose entity types are keyed by name
+///     rather than by CLR type, so several share one <see cref="System.Collections.Generic.Dictionary{TKey,TValue}" />.
+/// </summary>
+/// <remarks>
+///     Non-shared-model, so it goes through <see cref="NonSharedModelInfoCarrierHarness" /> (A49).
+///     EF's InMemory class adds a <c>ToInMemoryQuery</c> test of its own; that is the store's, not
+///     the spec base's, and is not carried (A47).
+/// </remarks>
+public class SharedTypeQueryInfoCarrierTest(NonSharedFixture fixture)
+    : SharedTypeQueryTestBase(fixture)
+{
+    private readonly NonSharedModelInfoCarrierHarness _harness = new(InfoCarrierTestStoreFactory.InMemory);
+
+    /// <inheritdoc />
+    protected override ITestStoreFactory TestStoreFactory
+        => _harness.TestStoreFactory;
+
+    /// <inheritdoc />
+    protected override ContextFactory<TContext> CreateContextFactory<TContext>(
+        Action<ModelBuilder>? onModelCreating = null,
+        Action<DbContextOptionsBuilder>? onConfiguring = null,
+        Func<IServiceCollection, IServiceCollection>? addServices = null,
+        Action<ModelConfigurationBuilder>? configureConventions = null,
+        Func<string, bool>? shouldLogCategory = null,
+        Func<TestStore>? createTestStore = null,
+        bool usePooling = true,
+        bool useServiceProvider = true)
+    {
+        Fixture = null;
+        _harness.Prepare(typeof(TContext), onModelCreating, addServices, onConfiguring, configureConventions, AddOptions);
+
+        return base.CreateContextFactory<TContext>(
+            onModelCreating, onConfiguring, addServices, configureConventions,
+            shouldLogCategory, createTestStore, usePooling, useServiceProvider);
+    }
+}
