+++ test/known-failures.names.txt
@@ -43,6 +43,8 @@ InfoCarrier.Core.FunctionalTests.Sqlite.Query.NullSemanticsQueryInfoCarrierTest.
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NullSemanticsQueryInfoCarrierTest.Switching_null_semantics_produces_different_cache_entry
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NullSemanticsQueryInfoCarrierTest.Where_contains_on_parameter_array_with_just_null_with_relational_null_semantics
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.NullSemanticsQueryInfoCarrierTest.Where_not_equal_using_relational_null_semantics_complex_in_equals
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.OwnedQueryInfoCarrierTest.Using_from_sql_on_owner_generates_join_with_table_for_owned_shared_dependents(async: False)
+InfoCarrier.Core.FunctionalTests.Sqlite.Query.OwnedQueryInfoCarrierTest.Using_from_sql_on_owner_generates_join_with_table_for_owned_shared_dependents(async: True)
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.PrimitiveCollectionsQuerySqliteInfoCarrierTest.Column_collection_equality_inline_collection_with_parameters
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.PrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_in_subquery_Union_another_parameter_collection_as_compiled_query
 InfoCarrier.Core.FunctionalTests.Sqlite.Query.PrimitiveCollectionsQuerySqliteInfoCarrierTest.Parameter_collection_in_subquery_and_Convert_as_compiled_query
